prevent browser zoom in presentation view
We had problems using this yesterday, because the people running the show mistakenly zoomed Chrome to 110%.
We should consider that all kind of people will run this, so I've added code to prevent zooming - the render()
calculation will consider the zoom ratio. It works quite nice actually.
Just set the constant to FALSE if you want to disable this feature. It is only tested in Chrome and Edge on Windows.